Rewrite the Gnus group activation method to be more efficient.
The main idea is to gather all groups based on their select methods,
and then issuing one `-request-list' per select method, instead of
activating them group by group.  This should in theory be much, much
faster, but there's a bunch of subtle interactions between a plethora
of variables that doesn't make this trivial to achieve.

This is a first stab at the issue, and seems to work, but I'm sure
that there are corner cases that doesn't quite work right.