(gnus-article-next-page): Change last-line-displayed behaviour.
When the last line of the article is displayed, scroll down once more
instead of going to the next article at once.

This makes for more SPC-ing, but since you normally don't know whether
an article is over or not in this situation, this is perhaps less
surprising and annoying behaviour.