Introduce gnus-registry-enabled and use it consistently.
* gnus-registry.el (gnus-registry-enabled): Add new variable.
(gnus-registry-install-shortcuts): Set `gnus-registry-install' to 'ask
only while we need to find out if it should be t or nil.
(gnus-registry-initialize): Don't set `gnus-registry-install' to t.
(gnus-registry-install-hooks): Set `gnus-registry-enabled' to t when
the registry is installed.  Set it to nil when it's unloaded.
(gnus-registry-install-p): Provide user guidance for the initial value
of `gnus-registry-install' when it's 'ask, otherwise return its value.
* nnregistry.el (nnregistry-open-server, nnregistry-server-opened): Use
`gnus-registry-enabled' instead of `gnus-registry-install'.
* nnmairix.el (nnmairix-determine-original-group-from-registry): Use
`gnus-registry-enabled' instead of `gnus-registry-install'.
* nnir.el (nnir-mode): Use `gnus-registry-enabled' instead of
`gnus-registry-install'.
From Dave Abrahams <dave@boostpro.com>.